pyproject.toml
[project]
name = "yosuki-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.96.0",
    "jsonschema>=4.26.0",
    "python-dotenv>=1.2.2",
    "replicate>=1.0.7",
]
